+++ test/lib/crawler/fetcher/retrier_test.exs
@@ -1,0 +1,7 @@
+defmodule Crawler.Fetcher.RetrierTest do
+  use Crawler.TestCase, async: true
+
+  alias Crawler.Fetcher.Retrier
+
+  doctest Retrier
+end